backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
motor==3.3.2
pymongo==4.6.0
google-cloud-bigquery==3.13.0
google-auth==2.25.2
requests==2.31.0
pandas==2.1.4
numpy==1.24.3
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
shopify-python-api==12.0.0
facebook-business==19.0.0
google-ads==23.1.0
aiofiles==23.2.1
celery==5.3.4
redis==5.0.1
